+++ pyproject.toml
@@ -38,7 +38,7 @@ markdown = "^3.1.1"
 markupsafe = "^2.0.1"
 
 paramiko = ">=2.6.0"
-prometheus-client = "^0.9.0"
+prometheus-client = "<1"
 psycopg2 = "^2.8.3"
 PyJWT = "^2.4.0"
 python_dateutil = "^2.6.1"
